requirements.txt
torch>=2.1.0
torchvision>=0.16.0
xgboost>=1.7.5
numpy>=1.24.0
pandas>=2.1.0
pillow>=10.0.0
pillow-heif>=0.9.3
optuna>=3.0.0
scikit-learn>=1.3.0
PyYAML>=6.0
